Add support for keeping defines that are also provided via command line
This patch adds support for a new way of handling defines that specified
both in the source code and on the command line, by allowing the source
code defines to remain unchanged.

Note that the both defines need to be the same in order for no warning
to be raised.