FIX: Failure to match #define parameter with -D symbol
This patch provides a new fix for case where the tool not matching a
`#define B  1` with a `-DB=1`, because it was considering all text after
the `#define B` as being the define value.

The above was happening because the chew state was being set as `PSEUDO_COMMENT`
before the call to `chew_on`. This seems to have been introduced in
version _Symbian v1.1.5_.